+++ src/managers/configuration.ts
@@ -57,8 +57,8 @@ export class ConfigurationManager implements IDisposable {
     const ruleWhitelist = config.get<string[]>("ruleWhitelist");
     const soundWhitelist = config.get<string[]>("soundWhitelist");
     const modWhitelist = config.get<string[]>("modWhitelist");
+    const prophecyWhitelist = config.get<string[]>("prophecyWhitelist");
     const performanceHints = config.get<boolean>("performanceHints");
-    const limitedModPool = config.get<boolean>("limitedModPool");
     const itemValueQuotes = config.get<boolean>("itemValueQuotes");
     const booleanQuotes = config.get<boolean>("booleanQuotes");
     const rarityQuotes = config.get<boolean>("rarityQuotes");
@@ -74,6 +74,7 @@ export class ConfigurationManager implements IDisposable {
       ruleWhitelist: ruleWhitelist ? ruleWhitelist : [],
       soundWhitelist: soundWhitelist ? soundWhitelist : [],
       modWhitelist: modWhitelist ? modWhitelist : [],
+      prophecyWhitelist: prophecyWhitelist ? prophecyWhitelist : [],
       performanceHints: performanceHints == null ? true : performanceHints,
       itemValueQuotes: itemValueQuotes == null ? true : itemValueQuotes,
       booleanQuotes: booleanQuotes == null ? false : booleanQuotes,